risk-analyzer/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
starlette==0.27.0

# Database connectors
psycopg2-binary==2.9.9
neo4j==5.15.0
qdrant-client==1.6.3
redis==5.0.0
pika==1.3.2
aio-pika==9.3.0
minio==7.1.17

# ML and data processing
numpy==1.24.3
pandas==2.0.3
sentence-transformers==2.2.2
nltk==3.8.1
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==10.1.0
pytesseract==0.3.10

# Utilities
pyjwt==2.8.0
python-dotenv==1.0.0
tenacity==8.2.3
httpx==0.25.1
loguru==0.7.2
aiofiles==23.2.1
colorama==0.4.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1 